move validation methods to VectorValidation
